Move console report
As first step we need to move the console report away of the status
reporter class.

Related https://github.com/fastruby/skunk/issues/50
