+++ .github/workflows/main.yml
@@ -1,6 +1,7 @@
 # .github/workflows/main.yml
 
 name: CI
+
 on: [push, pull_request]
 
 jobs:
