+++ .rubocop_todo.yml
@@ -1,30 +1,29 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2021-08-31 01:33:08 UTC using RuboCop version 1.9.1.
+# on 2023-02-07 02:28:48 UTC using RuboCop version 1.24.1.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
 # versions of RuboCop, may require this file to be generated again.
 
-# Offense count: 2
+# Offense count: 1
 # Configuration parameters: Include.
 # Include: **/*.gemspec
 Gemspec/RequiredRubyVersion:
   Exclude:
-    - 'gemfiles/skunk.gemspec'
     - 'skunk.gemspec'
 
 # Offense count: 1
 # Cop supports --auto-correct.
 Layout/ClosingHeredocIndentation:
   Exclude:
-    - 'lib/skunk/cli/commands/status_reporter.rb'
+    - 'lib/skunk/cli/generators/console_report.rb'
 
 # Offense count: 1
 # Cop supports --auto-correct.
 Layout/HeredocIndentation:
   Exclude:
-    - 'lib/skunk/cli/commands/status_reporter.rb'
+    - 'lib/skunk/cli/generators/console_report.rb'
 
 # Offense count: 2
 Lint/MissingSuper:
